Refactor data download API into separate category
