+++ stroom-data/stroom-data-store-api/build.gradle
@@ -3,6 +3,10 @@ ext.moduleName = 'stroom.data.store.api'
 dependencies {
     implementation project(':stroom-core-shared')
     implementation project(':stroom-meta:stroom-meta-api')
+    implementation project(':stroom-util-shared')
 
     implementation libs.jackson_annotations
+    implementation libs.javax_inject
+    implementation libs.swagger_annotations
+    implementation libs.ws_rs_api
 }
