+++ stroom-data/stroom-data-store-impl/src/main/java/stroom/data/store/impl/DataResourceImpl.java
@@ -29,9 +29,7 @@ import stroom.pipeline.shared.AbstractFetchDataResult;
 import stroom.pipeline.shared.FetchDataRequest;
 import stroom.pipeline.shared.FetchDataRequest.DisplayMode;
 import stroom.pipeline.shared.FetchDataResult;
-import stroom.resource.api.ResourceStore;
 import stroom.util.EntityServiceExceptionUtil;
-import stroom.util.io.StreamUtil;
 import stroom.util.shared.Count;
 import stroom.util.shared.FetchWithLongId;
 import stroom.util.shared.OffsetRange;
@@ -40,39 +38,27 @@ import stroom.util.shared.ResourceKey;
 
 import event.logging.ComplexLoggedOutcome;
 import event.logging.ExportEventAction;
-import event.logging.File;
 import event.logging.ImportEventAction;
 import event.logging.MultiObject;
 import event.logging.ViewEventAction;
 
-import java.io.IOException;
-import java.io.InputStream;
-import java.math.BigInteger;
-import java.nio.file.Files;
-import java.nio.file.Path;
 import java.util.Collections;
 import java.util.List;
 import java.util.Set;
 import javax.inject.Inject;
 import javax.inject.Provider;
-import javax.ws.rs.core.MediaType;
-import javax.ws.rs.core.Response;
-import javax.ws.rs.core.StreamingOutput;
 
 @AutoLogged
 class DataResourceImpl implements DataResource, FetchWithLongId<List<DataInfoSection>> {
 
     private final Provider<DataService> dataServiceProvider;
     private final Provider<StroomEventLoggingService> stroomEventLoggingServiceProvider;
-    private final Provider<ResourceStore> resourceStoreProvider;
 
     @Inject
     DataResourceImpl(final Provider<DataService> dataServiceProvider,
-                     final Provider<StroomEventLoggingService> stroomEventLoggingServiceProvider,
-                     final Provider<ResourceStore> resourceStoreProvider) {
+                     final Provider<StroomEventLoggingService> stroomEventLoggingServiceProvider) {
         this.dataServiceProvider = dataServiceProvider;
         this.stroomEventLoggingServiceProvider = stroomEventLoggingServiceProvider;
-        this.resourceStoreProvider = resourceStoreProvider;
     }
 
     @AutoLogged(OperationType.MANUALLY_LOGGED)
@@ -102,55 +88,6 @@ class DataResourceImpl implements DataResource, FetchWithLongId<List<DataInfoSec
                     }
                 })
                 .getResultAndLog();
-    }
-
-    @AutoLogged(OperationType.MANUALLY_LOGGED)
-    @Override
-    public Response downloadZip(final FindMetaCriteria criteria) {
-
-        final ResourceGeneration resourceGeneration = dataServiceProvider.get().download(criteria);
-        final ResourceStore resourceStore = resourceStoreProvider.get();
-        final ResourceKey resourceKey = resourceGeneration.getResourceKey();
-        final Path tempFile = resourceStore.getTempFile(resourceKey);
-
-        try {
-            final ExportEventAction exportEventAction = ExportEventAction.builder()
-                    .withSource(MultiObject.builder()
-                            .addCriteria(stroomEventLoggingServiceProvider.get().convertExpressionCriteria(
-                                    "Meta",
-                                    criteria))
-                            .addFile(File.builder()
-                                    .withName(tempFile.getFileName().toString())
-                                    .withSize(BigInteger.valueOf(Files.size(tempFile)))
-                                    .build())
-                            .build())
-                    .build();
-
-            return stroomEventLoggingServiceProvider.get()
-                    .loggedWorkBuilder()
-                    .withTypeId(StroomEventLoggingUtil.buildTypeId(this, "downloadZip"))
-                    .withDescription("Downloading stream data as zip")
-                    .withDefaultEventAction(exportEventAction)
-                    .withSimpleLoggedResult(() -> {
-                        // Stream the downloaded content to the client as ZIP data
-                        final StreamingOutput streamingOutput = output -> {
-                            try (final InputStream is = Files.newInputStream(tempFile)) {
-                                StreamUtil.streamToStream(is, output);
-                            } finally {
-                                resourceStore.deleteTempFile(resourceKey);
-                            }
-                        };
-
-                        return Response
-                                .ok(streamingOutput, MediaType.APPLICATION_OCTET_STREAM)
-                                .header("Content-Disposition", "attachment; filename=\"" +
-                                        tempFile.getFileName().toString() + "\"")
-                                .build();
-                    })
-                    .getResultAndLog();
-        } catch (IOException e) {
-            throw EntityServiceExceptionUtil.create(e);
-        }
     }
 
     @AutoLogged(OperationType.MANUALLY_LOGGED)
